edid-decode: output full frequencies for 4:2:0
Don't half hfreq for 4:2:0 timings - the character clock is halved, but the
number of characters per line is also halved (two luminance values per
character for 4:2:0), so there's no change in hfreq.

Don't half pixel clock because it looks weird. Character clock is halved
but the number of pixels remains the same (two luminance values per
character for 4:2:0).

Continue to use the half hfreq and half pixel clock for the ranges
calculations because some non-HDMI 2.0 displays have max pixel clock halved
when they have 4:2:0 modes and because other reasons that lead to adding
this code in the first place?

Signed-off-by: Joe van Tunen <joevt@shaw.ca>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>