edid-decode: allow undefined aspect ratio
Distinguish between "undefined" and "reserved" aspect ratios. See examples
apple-xdr-6k-tile0, lg-ultrafine-5k-v1-thunderbolt-dp1-tile0 and
asus-xg438q-dp.

"undefined" is for an aspect ratio that doesn't exist in the list of known
aspect ratios. This does not deserve a fail.

"reserved" should not be used and still deserves the fail.

Signed-off-by: Joe van Tunen <joevt@shaw.ca>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>