edid-decode: convert Dolby Vision PQ values to cd/m^2
Convert the Dolby Vision PQ values to cd/m^2. Since I do not
have the DV spec, I made an educated guess what the divider is
to map the integer PQ value to a value in the range 0-1 which
is needed as input to the pq2nits function.

It should be pretty close to the actual value, though.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>