edid-decode: fix standard timing vertical pixels
Don't do ceiling to nearest 8 pixels for active vertical lines.
See examples elo-4600l-hdmi and kogan-kaled24144f-hdmi.

Section 3.9 and 3.10.3.6 of EDID 1.4 does not say vertical lines must be
a multiple of 8.  This line of code appears to have been added to satisfy
the 3rd example in VTB-EXT spec but that example has an incorrect HAP
indicator decimal value so it cannot be trusted. Also, all 3 examples have
an incorrect vertical refresh value as noted in parse-vtb-ext-block.cpp.
The VESA DMT spec has the following examples that are not a multiple of 8
lines which support this change:

1400x1050 4:3
1440x900 16:10
1600x900 16:9
1680x1050 16:10

Finally, Ref. D-8 of EDID 1.4 says about Section 3.9 that "The Standard
Timings Identification code may not be used to identify timings which do
not match one of these standard aspect ratios."

If vertical lines is odd then a warning is output. That way an attempt to
use ST to describe 1360x768 (a common resolution) will result in a warning
(since the nearest result that can be described by an ST is 1360x765).

Signed-off-by: Joe van Tunen <joevt@shaw.ca>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>