edid-decode: Output block type before fail
For extended tags, block type should be output before fail messages
(duplicate failure, or missing audio failure).
For normal tags 0x04 and 0x05, fail message should appear after block type
instead of block data to be consistent (and also to indicate that the
failure is because of the block type and not the contents of the block).

Signed-off-by: Joe van Tunen <joevt@shaw.ca>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>