edid-decode: move unknown block warning
Also, make cta_block like cta_ext_block

First, fix the first switch statement in cta_ext_block. The first switch
statement sets or clears data_block (the name of the block which is used in the
Warnings and Failures section of the output).  Remove hex_block because it will
be done by the third switch statement that handles outputting the contents of
each block.  The "Unknown Extended CTA-861 Data Block 0x%02x.\n" warning now
appears after the block name instead of after the block contents (because it's
not the contents that caused the error).

Then, in cta_block:
1) Create a first switch statement like that of cta_ext_block (described
   above).  The "Unknown CTA-861 Data Block %u.\n" warning now appears after
   the block name instead of after the block contents (because it's not the
   contents that caused the error).
2) Create a second switch statement like that of cta_ext_block. It handles
   checking for duplicate blocks.
3) After checking for duplicates, check cta_byte3 / audio_block discrepancy,
   exactly like cta_ext_block does. These lines come from after the original
   switch statement of cta_block.
4) The original switch statement of cta_block has lines that are moved to the
   new first and second switch statements.

Signed-off-by: Joe van Tunen <joevt@shaw.ca>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>