edid-decode: DisplayID non-0 filler fixes
It's probably not a Product Identification Data Block or any kind of block so
data_block should be cleared (so that the block name doesn't appear in the
Failures section for this fail).  Output hex data because it might contain
interesting data (it's at least known to be not zero).

Signed-off-by: Joe van Tunen <joevt@shaw.ca>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>