edid-decode: output correct unknown block length
The correct length of a DisplayID block is len. It is the number of bytes in
the payload bytes (doesn't include the tag or length or revision). It may
include the OUI but we don't output the OUI in hex_block, and an unknown block
can't have an OUI unless there's a new DisplayID spec with a new type of block
that we don't know about.

The length parameter passed to displayid_block is the number of bytes that
remain in a DisplayID extension block (maybe it needs to be renamed?). We pass
it to displayid_block so it can handle length related errors. displayid_block
handles all interpretation of the data block. The only byte we know that can be
interpreted on entry is the tag byte because it is the first byte and length is
at least one (according to the for loop in parse_displayid_block).  The payload
length byte occurs after the tag byte - it might exist beyond the limit of
length. If that's the case, we can at least report the tag block type with the
error, but we choose not to. Instead, we output the rest of the bytes
(including the tag) as hex because it probably isn't a real block.

Signed-off-by: Joe van Tunen <joevt@shaw.ca>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>