edid-decode: issue warning if sRGB support isn't indicated
RGB colorimetry is ambiguous: it can be interpreted as either sRGB
(Windows and linux display drivers assume sRGB) or as using the
default RGB colorimetry as defined in the base block (Mac drivers
do this).

This is not a problem if the base colorimetry is equal to sRGB, but if
it differs, then it is recommended to indicate sRGB explicitly in the
Colorimetry Data Block. Support for this bit was added in CTA-861.6.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>