edid-decode: add --replace-serial-numbers option
This replaces any serial numbers by '123456' in the EDID itself.
It differs from the --hide-serial-numbers option that just shows
them as '...' in the human readable output, but leaves them in the
EDID hex dump.

Replacing the serial numbers will also update the checksums, so
incorrect checksums in the original EDID will no longer be detected
when using this option.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>