edid-decode: check if DTD == HDMI 2 but VICs HDMI 1
If the DTD matches a VIC with pixelclock > 340 MHz, and
the highest VIC pixelclock <= 340 MHz, then report a
mismatch: DTD is an HDMI 2.x timing, while all VICs
are HDMI 1.x timings.

This suggests that the display has HDMI 2.x support disabled
(a common menu setting), and that HDMI 2.x VICs are removed
from the CTA Extension Block, but the DTD in the Base Block
is not updated and still lists an HDMI 2.x timing.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>