edid-decode: add initial support for parsing InfoFrames
This adds initial support for parsing InfoFrames.

Starting with kernel v6.11-rc1 InfoFrames are exposed in debugfs
by some HDMI output drivers (Raspberry Pi among them).

These can be specified with the new --infoframe (or -I) option.
This option can be used multiple times to specify multiple
InfoFrames.

Each InfoFrame is parsed by edid-decode, and it is also checked
for conformity. Note that the conformity checks are still
work-in-progress for the AVI and HDMI InfoFrames.

It can also be combined with specifying the EDID of the display
the InfoFrames are sent to, and that will allow for more strict
conformity checks against that EDID.

This is also still work-in-progress.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>