test/vesa-displayid-fixed-length.test: add test example
Add a Fixed-length DisplayID example from the DisplayID
2.1a standard.

I've never seen this EDID replacement in the wild, the only
one I ever saw was this example in the standard.

Note that edid-decode can't (yet?) handle this. But at least
it is there if I want to add support for this format.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>