edid-decode: the AMD VSDB minor version is really a caps field
This kernel commit:

ec8e59cb4e0c ("drm/amd/display: Get replay info from VSDB")

gives a bit of insight into the AMD VSDB. Specifically, the first
byte is the version and the second byte is a Feature Caps byte.
Only one capability was described in that commit, so I added that.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>